+++ Lib/Utils/CommandLineParser/Definitions/FindUnusedCommand.cs
@@ -1,0 +1,9 @@
+namespace Lib.Utils.CommandLineParser.Definitions
+{
+    public class FindUnusedCommand : CommonParametersBaseCommand
+    {
+        public override string[] Words => new[] { "findunused", "fu" };
+
+        protected override string Description => "find unused TypeScript sources";
+    }
+}